+++ .vscode/settings.json
@@ -1,3 +1,4 @@
 {
     "rust-analyzer.cargo.allTargets": false,
+    "rust-analyzer.cargo.features": "all",
 }
