+++ Cargo.toml
@@ -14,4 +14,14 @@ categories = ["embedded", "no-std"]
 
 [dependencies]
 defmt = "0.3.8"
-embedded-hal = "1.0.0"
+embedded-hal = { version = "1.0.0", optional = true } 
+embedded-hal-async = { version = "1.0.0", optional = true } 
+
+[features]
+default = ["sync-driver"]
+sync-driver = ["embedded-hal"]
+async-driver = ["embedded-hal-async"]
+
+
+[package.metadata.docs.rs]
+all-features = true
